(core/lua/platform) namespace refactoring
A long list of changes and first round of documentation changes to
reflect on the new ns:/ prefix for resources. The db- to platform lookup
side is ok, it is the lua->platform mapping that still needs some
experiments to get further.

The main remaining bit is to make sure that the different 'rwp' mappings
are actually respected so that write will fail on an explicit userns
reference for write.